integer overflow with avr8

+++ include/rtc.h
@@ -25,10 +25,10 @@ public:
 	Date(unsigned int Y, unsigned int M, unsigned int D,
 	     unsigned int h, unsigned int m, unsigned int s) 
 	    : _Y(Y), _M(M), _D(D), _h(h), _m(m), _s(s) {}
-	Date(const Second & seconds, unsigned int epoch_days = 0);
+	Date(const Second & seconds, unsigned long epoch_days = 0);
  
 	operator Second() const { return to_offset(); }
-	Second to_offset(unsigned int epoch_days = 0) const;
+	Second to_offset(unsigned long epoch_days = 0) const;
 
 	unsigned int year() const { return _Y; };
 	unsigned int month() const { return _M; };
